[sui archival] finish removing usage
