--- crates/sui-archival/Cargo.toml
@@ -1,40 +1,0 @@
-[package]
-name = "sui-archival"
-version.workspace = true
-edition = "2021"
-publish = false
-license = "Apache-2.0"
-authors = ["Mysten Labs <build@mystenlabs.com>"]
-
-[dependencies]
-indicatif.workspace = true
-anyhow.workspace = true
-serde.workspace = true
-byteorder.workspace = true
-tracing.workspace = true
-bytes.workspace = true
-num_enum.workspace = true
-futures.workspace = true
-rand.workspace = true
-object_store.workspace = true
-prometheus.workspace = true
-sui-config.workspace = true
-sui-types.workspace = true
-sui-storage.workspace = true
-fastcrypto = { workspace = true, features = ["copy_key"] }
-tokio = { workspace = true, features = ["full"] }
-serde_json = "1.0.95"
-
-[dev-dependencies]
-tempfile.workspace = true
-more-asserts.workspace = true
-move-binary-format.workspace = true
-move-core-types.workspace = true
-tokio = { workspace = true, features = ["test-util"] }
-ed25519-consensus.workspace = true
-fastcrypto.workspace = true
-sui-swarm-config.workspace = true
-sui-macros.workspace = true
-
-[target.'cfg(msim)'.dependencies]
-sui-simulator.workspace = true
