+++ crates/sui-data-ingestion/Cargo.toml
@@ -26,7 +26,6 @@ prometheus.workspace = true
 telemetry-subscribers.workspace = true
 tokio = { workspace = true, features = ["full"] }
 tracing.workspace = true
-sui-archival.workspace = true
 sui-storage.workspace = true
 sui-data-ingestion-core.workspace = true
 sui-kvstore.workspace = true
