+++ crates/sui-data-ingestion/src/lib.rs
@@ -5,4 +5,4 @@ mod progress_store;
 mod workers;
 
 pub use progress_store::DynamoDBProgressStore;
-pub use workers::{ArchivalConfig, ArchivalReducer, ArchivalWorker, BlobTaskConfig, BlobWorker};
+pub use workers::{BlobTaskConfig, BlobWorker};
