+++ crates/sui-data-ingestion/src/workers/mod.rs
@@ -1,7 +1,5 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-mod archival;
 mod blob;
-pub use archival::{ArchivalConfig, ArchivalReducer, ArchivalWorker};
 pub use blob::{BlobTaskConfig, BlobWorker};
