+++ scripts/simtest/simtest-run.sh
@@ -52,7 +52,6 @@ scripts/simtest/cargo-simtest simtest \
   --color always \
   --test-threads "$NUM_CPUS" \
   --package sui-core \
-  --package sui-archival \
   --package sui-e2e-tests \
   --profile simtestnightly \
   -E "$TEST_FILTER" 2>&1 | tee "$LOG_FILE"
